move test suite to testutil
